moved context_utils.py
